pyproject.toml
[project]
name = "atlas"
version = "0.1.0"
description = "Saudi AI Middleware - Enterprise AI orchestration platform"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "oracledb>=2.0.0",
    "sentence-transformers>=2.2.0",
    "qdrant-client>=1.7.0",
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.0.0",
    "pydantic[email]>=2.0.0",
    "bcrypt>=4.0.0",
    "PyJWT>=2.8.0",
    "PyYAML>=6.0",
    "numpy>=1.26.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/atlas"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
pythonpath = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
